Rename directory `Basic` to `basic`
Rationale: ROS does not define a naming convention for directory names
in packages, but in general it seems that they are often named according
to the naming conventions for identifiers in ROS package names, C-family
programming languages, or CMake conventions. These are (as far as I
know) lower_snake_case, so I rename them as such.

Signed-off-by: yamacir-kit <httperror@404-notfound.jp>